Fixed text position
Now text is centered, and also "30" is inside the canvas and a little bigger. Also text is in front of "30", but behind confetti.
Also, last commit I forgot to say that I also added the confetti rain when the device is shaken (hope it works cause i didn't test it yet).